condaのis_init対応

+++ scripts/conda.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 set -xe
 
-curl https://repo.anaconda.com/archive/Anaconda3-2022.05-Linux-x86_64.sh -o /tmp/conda.sh
-source ${HOME}/anaconda3/etc/profile.d/conda.sh
+if [[ "$is_init" = true ]]; then
+    rm -rf ${HOME}/anaconda3
+fi
 
-conda update conda
-conda install anaconda
+if [[ ! -e "$HOME/anaconda3" ]]; then
+    curl https://repo.anaconda.com/archive/Anaconda3-2022.05-Linux-x86_64.sh -o /tmp/conda.sh
+    sh /tmp/conda.sh -b
+    source ${HOME}/anaconda3/etc/profile.d/conda.sh
+
+    conda update conda
+    conda install anaconda
+fi
+
+
